more instruction implementing and misc changes

+++ byteutils.go
@@ -10,8 +10,8 @@ func get_nnn(opcode uint16) uint16 {
 	return opcode & 0x0FFF
 }
 
-func get_nibble(opcode uint16) uint16 {
-	return opcode & 0x000F
+func get_nibble(opcode uint16) byte {
+	return byte(opcode & 0x000F)
 }
 
 func get_x(opcode uint16) byte {
@@ -26,6 +26,6 @@ func get_low_byte(opcode uint16) byte {
 	return byte(opcode & 0x00FF)
 }
 
-func get_high_nibble(opcode uint16) uint16 {
-	return (opcode & 0xF000) >> 12
+func get_high_nibble(opcode uint16) byte {
+	return byte((opcode & 0xF000) >> 12)
 }
